Ouin ouin Alexis a casté en int un calcul qui nécessitait un float FreeSurface des persos juste après leur création Ratio proportionnel des positions des joueurs

+++ include/personnages.h
@@ -9,7 +9,7 @@
 
 /* struct qui permet de faire le menu de selection des persos */
 typedef struct {
-  int w, h;
+  float w, h;
 } taille;
 
 typedef struct {
